phase-03/backend/requirements-dev.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
pydantic==2.9.2
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
SQLAlchemy==2.0.35
alembic==1.13.3
pytest==8.3.3
httpx==0.27.2